+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - Space for the cache file is now pre-allocated before SCIM operations start (#197)
   
 #### Bugfixes
+  - Fixed an access violation which could occur when generating activities if no groups had been loaded (#208)
   - When using the --rebuild-cache feature, a failure to GET objects will now terminate the client (#205)
 
 ## v2.18 (2024-11-19)